# Conflicts:
#	.gitignore
#	WOT-iOS/WOTData/WOTData.xcodeproj/xcuserdata/pavel_yeshchyk.xcuserdatad/xcschemes/xcschememanagement.plist
#	WOT-iOS/WOTPivot/WOTPivot.xcodeproj/xcuserdata/pavel_yeshchyk.xcuserdatad/xcschemes/xcschememanagement.plist

+++ WOT-iOS/WOT-iOS/ui/WOTTankPivotViewController/cells/WOTTankPivotDataCollectionViewCell.m
@@ -48,7 +48,7 @@
 }
 
 - (void)setUuid:(NSInteger)uuid {
-    self.dataLabel.text = [NSString stringWithFormat:@"%d",uuid];
+    self.dataLabel.text = [NSString stringWithFormat:@"%ld",(long)uuid];
 }
 
 - (void)setMask:(NSString *)mask {
